dtp ps: implemented NULL <--> default support
This also includes some reorganization of default policies code.
