+++ linux/net/rina/dtp-ps-default.h
@@ -1,15 +1,14 @@
 /*
- * Common function for DTP's policy set
+ * Default policy set for DTP
  *
  *    Vincenzo Maffione <v.maffione@nextworks.it>
  *    Francesco Salvestrini <f.salvestrini@nextworks.it>
  *    Leonardo Bergesio <leonardo.bergesio@i2cat.net>
- *    Miquel Tarzan <miquel.tarzan@i2cat.net>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation; either version 2 of the License, or
- * (at your option); any later version.
+ * (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -21,32 +20,23 @@
  * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  */
 
-#ifndef RINA_DTP_PS_COMMON_H
-#define RINA_DTP_PS_COMMON_H
+#ifndef DTP_PS_DEFAULT_H
+#define DTP_PS_DEFAULT_H
 
 #include "dtp-ps.h"
-#include "dtcp.h"
+#include "dtp.h"
 
-int
-common_transmission_control(struct dtp_ps * ps, struct pdu * pdu);
+int default_transmission_control(struct dtp_ps * ps, struct pdu * pdu);
 
-int
-common_closed_window(struct dtp_ps * ps, struct pdu * pdu);
+int default_closed_window(struct dtp_ps * ps, struct pdu * pdu);
 
-int
-common_flow_control_overrun(struct dtp_ps * ps,
-                            struct pdu *    pdu);
+int default_flow_control_overrun(struct dtp_ps * ps,
+                                 struct pdu *    pdu);
 
-int
-common_initial_sequence_number(struct dtp_ps * ps);
+int default_initial_sequence_number(struct dtp_ps * ps);
 
-int
-common_receiver_inactivity_timer(struct dtp_ps * ps);
+int default_receiver_inactivity_timer(struct dtp_ps * ps);
 
-int
-common_sender_inactivity_timer(struct dtp_ps * ps);
+int default_sender_inactivity_timer(struct dtp_ps * ps);
 
-int dtp_ps_common_set_policy_set_param(struct ps_base * bps,
-                                       const char    * name,
-                                       const char    * value);
-#endif
+#endif /* DTP_PS_DEFAULT_H */
